+++ sydtest/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## Unreleased: [0.8.0.0]
+
+### Changed
+
+* The `TestDefM` now contains a `TestDefEnv` which also contains the test description path, along with `TestRunSettings`.
+  It's unlikely that you depended on this, but requires a major version bump nonetheless.
+
+### Added
+
+* `getTestDescriptionPath` to get the test description path upwards from inside a test definition.
+
 ## [0.7.0.1] - 2021-12-23
 
 ### Changed